codegen: use templates (#4)
managing the print statements for each line or chunk of generated code was getting quite out of hand. this change uses `text/template` for each component of the generated source file. it brings the release binary size up to ~2.6M (from ~1.8M), but it massively improves the overall code quality.