refactor: only write the output file if there was no generator error (#10)
previously, if there was an error while generating, the output file would
contain partial code output, causing followup compiler errors for the project
of which it's a part. this change makes it so that the generator writes to a
buffer, and then the buffer is copied to the output file.